authenticated routes

+++ app/routes/teams/index.js
@@ -3,8 +3,10 @@ import Ember from 'ember';
 export default Ember.Route.extend({
   actions: {
     remove: function(model) {
-      if (confirm('Are you sure?')) {
-        model.destroyRecord();
+      if (this.get('session.isAuthenticated')) {
+        if (confirm('Are you sure?')) {
+          model.destroyRecord();
+        }
       }
     }
   },
